move admin into client, add basic router
