--- admin/.env.example
@@ -1,1 +1,0 @@
-REACT_APP_SERVER_URL=http://localhost:3000
