--- admin/README.md
@@ -1,3 +1,0 @@
-# opensourceradio admin
-
-react app for managing schedules and the asset library
