--- admin/package.json
@@ -1,33 +1,0 @@
-{
-  "name": "admin",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "@dmamills/stylish": "^1.2.2",
-    "classname": "0.0.0",
-    "classnames": "^2.2.6",
-    "dropzone": "^5.5.1",
-    "moment": "^2.24.0",
-    "react": "^16.8.6",
-    "react-datepicker": "^2.3.0",
-    "react-dom": "^16.8.6",
-    "react-router-dom": "^5.0.0",
-    "react-scripts": "2.1.8",
-    "react-select": "^2.4.2"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": "react-app"
-  },
-  "browserslist": [
-    ">0.2%",
-    "not dead",
-    "not ie <= 11",
-    "not op_mini all"
-  ]
-}
