+++ client/package.json
@@ -7,11 +7,15 @@
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.4.0",
     "classnames": "^2.2.6",
+    "dropzone": "^5.5.1",
     "flv.js": "^1.5.0",
     "moment": "^2.24.0",
     "react": "^16.12.0",
+    "react-datepicker": "^2.10.1",
     "react-dom": "^16.12.0",
+    "react-router-dom": "^5.1.2",
     "react-scripts": "2.1.5",
+    "react-select": "^2.4.2",
     "socket.io": "^2.2.0",
     "socket.io-client": "^2.2.0"
   },
