+++ client/src/admin/Library/Browser/Folder.js
